Add a button to allow users resize the grid
Users can click on the button to enter a custom value for the number of
cells they want on the grid.
Only numbers within the range of 0 and 100 will be accepted.
Floats will be floored to the nearest integer values.
If a user enters a negative number, it's absolute value will be used
instead.